made coverage better also im sorry for using global you can choose not to merge this because of my sin
